Hack to have ffmpeg's config.log in build64
Attempt to follow a shinchiro change. Hopefully this is idempotent.
Because mpv-winbuild sets `ALWAYS_REMOVE_BUILDFILES`, it gets deleted.
This attempts to patch ffmpeg.cmake to have the log created in `$builddir`
While I could modify my own mpv-winbuild-cmake directly, it is this
that sets the deletion flag. Plus, this is a bonus: I'm not too bothered
if this doesn't apply, so I don't want to add extra maintenance work.